+++ CHANGELOG.md
@@ -4,11 +4,17 @@
 
 ## [Unreleased]
 
-## [0.0.1] - 2022-11-09
+### Added
+
+- Added an option to disable the plugin for large files.
+- Added service optimizations for performance gains.
+
+## [0.0.1] - 2024-11-09
 
 ### Added
 
 - Initial release
 
 [Unreleased]: https://github.com/j-d-ha/brackets/compare/v0.0.1...HEAD
+
 [0.0.1]: https://github.com/j-d-ha/brackets/commits/v0.0.1